change expected request shape for `/query` and `/history` to use `from
Instead of outer keys specifying `query` and `ledger` for these requests, we now expect just the queries themselves, with `from` identifying the ledger, eg

```
{
  "from": "fluree/cookbook",
  "select": {"?s": ["*"]},
  "where": [
    ["?s", "ex:follows", "ex:cam"]
  ]
}
```

For now, the addition of this key is contained just to the http-api layer. This means that the schema for http-api-gateway and that of the inner db library fns are different, for the time being. To accommodate this, we use the new fns in db that allow us to amend query schema with key/value pairs. Eventually, we will support `from` at the db lib layer, and we won't need this anymore.